Renamed the file strong_types.h to file_path.h
