+++ Eclipse/src/eclipse/renderer/shader.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "eclipse/core/core.h"
-#include "eclipse/common_types/strong_types.h"
+#include "eclipse/common_types/file_path.h"
 
 #include <glm/glm.hpp>
